audiosystem: fix a huge performance issue caused by the recent change to CalculateTime
When the background music finished playing for the first time, it would then bug out, and BASS would cause huge file system reads, dropping the FPS by a huge amount.
This was because it would try to set the time for the channel since it thought the music went out of sync, but the time to sync it was calculated wrong, causing it to constantly try to set the time, which then caused BASS to read the entire file over and over again.