gamemode: properly handle player disconnects while waiting for everyone to join.
After 90 seconds it will try to force start the round, if the slasher left, all players are asked if they want to possibly become the slasher, and they get 15 seconds to decide, if they don't pick any option it assumed they don't want to be a slasher.
If all players decline, it will return to the lobby since it's missing a slasher.
If at least one player accepts (if multiple accept, one is randomly picked) they will be given the option to select the slasher with the same options as they had in the lobby, he gets 15 seconds to decide.
If he doesn't pick any slasher in the 15 seconds, he will be given the slasher that was originally selected by the previous player that left.
If he disconnects while he's supposed to be selecting a slasher, then it will pick another random player that had agreed to become a slasher and they're given the same option. If no players are left to randomly pick from it will return to the lobby.

This should hopefully ensure that every possible scenario is properly handled.