slashers: do very basic balancement based on player count that probably will screw up a lot
Yes, this will most likely cause a lot of issues, though with feedback it should be somewhat easy to get this into a good state.
I don't have that many people to test it out, so this basically is just many random values hoping to not break stuff.
It can also be reverted if it becomes too big of an issue, but it's better than nothing.
Stuff like more tasks/generators/minigames can be added at a later point.