gamemode: multiple changes
[+] Allow the second slasher in duality to pick a slasher too
[+] GMod will now flash in the Windows taskbar if you have to select a slasher or the round is about to start
[+] Half of the earned points are counted as experience
[#] Remove the hardcoded 2-slasher things
In theory, a server could now force more slashers simply by adding them to the AssignedSlashers list
[#] Fixed slasher show tick thingy using a hardcoded value instead of FrameTime
[#] Changed networking of LobbySlasherInformation
[#] Changed how SlashCo.GetRandomSlasher works to be more efficient and to account for dangerlevel & class
This also solves that in duality the second slasher could have a different dangerlevel or class than intended.
This also applies to Nightmare!
[#] Fixed HUD rendering not being called when having the main menu open
We now use the hook SlashCo:DrawHUD instead of GMod's DrawHUD, as ours is more reliable.
This fixes a lot of things not being rendered while having the main menu open
[#] Fixed SlasherHUD sometimes not being removed if a slasher is switched to another server
This normally cannot even happen, though if a server decides to switch a slasher to another slasher, this may have happened.
[#] Rename some fields for internal consistency
It was pissing me off that most tables used SteamID, but then some others were like s_id or just id
[#] Fixed slasher selection not rendering when the main menu was open
[#] Fixed round points being lost on autorefresh
[#] Fixed background music playing in roundover screen